requirements.txt
pylint
yapf
django
pylint-django
djangorestframework
djangorestframework-simplejwt
selenium